misc: Add negation dependency of QCOM_SI_CORE feature to QCOM_SMCI_PROXY
The SMCI-Proxy driver proxies requests from in-tree kernel clients to
out-of-tree SMCInvoke driver. However, when the in-tree SI-core driver
is enabled, the SMCI-Proxy driver is not required since clients can
directly call the same APIs exported by the SI-core driver. This change
ensures that the two drivers are always enabled mutually exclusively in
the build via menuconfig.

Change-Id: Ia1ea3a1330f0b2f32a7aa005e747ef0f699fb430
Signed-off-by: Harshal Dev <quic_hdev@quicinc.com>