msm: ep_pcie: Fix issue with parsing clock frequencies
Use the clock frequencies provided in the device tree only
when they are explicitly specified. Otherwise, fall back to
the statically mapped clock frequencies.

We are trying to access the buffer where the device-tree
supplied frequencies are copied, even when those frequencies
are not provided. This leads to an out-of-bounds array access.

In the static clock table, set the pipe clock frequency to zero
so that it can be configured by the controller. Manually setting
the pipe clock frequency resulting in linkup failure.

Change-Id: Ieb493419e0dbb6d326574d4a4344aafa77abbb81
Signed-off-by: Veerabhadrarao Badiganti <vbadigan@qti.qualcomm.com>